This commit was manufactured by cvs2svn to create tag 'release-3-2-1'.

+++ source/common/unistr_cnv.cpp
@@ -267,7 +267,7 @@ UnicodeString::doCodepageCreate(const char *codepageData,
     return;
   }
   if(dataLength == -1) {
-    dataLength = (int32_t)uprv_strlen(codepageData);
+    dataLength = uprv_strlen(codepageData);
   }
 
   UErrorCode status = U_ZERO_ERROR;
